move update settings to updates tab
